pyproject.toml
[project]
name = "sunset"
version = "0.1.0"
description = "SUnSET: Timeline Summarization with Stakeholder-Event-Time triplets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LLM
    "openai>=1.0.0",

    # Embedding & ML
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",  # For local GTE embedding

    # NLP
    "spacy>=3.7.0",

    # Evaluation
    "tilse>=0.1.0",
    "rouge-score>=0.1.2",

    # HTTP & Async
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",

    # Data
    "tqdm>=4.66.0",

    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
